Support deadline override in cassandra schema job
With the default `activeDeadlineSeconds` of 120, in our cluster the jaeger-cassandra pods may not be available quickly enough for the job to complete. After its 120 second deadline, the job is aborted and will not be retried. We would like to increase this deadline in our configuration so that this chart may fully install without manual intervention.